requirements.txt
asgiref==3.2.3
Django==3.0.2
django-filter==2.2.0
django-mysql==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.1.1
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
